+++ src/BF2TV.Frontend/Pages/FriendList.razor
@@ -2,6 +2,7 @@
 @using BF2TV.Frontend.Pages.Friends
 @using BF2TV.Frontend.Store.FriendList
 @using BF2TV.Frontend.Store
+@using BF2TV.Domain.Models
 @inherits FluxorComponent
 @inject IDispatcher _dispatcher
 @inject IState<FriendListState> _friendListState
