+++ src/BF2TV.Frontend/Pages/Friends/ToggleFriendIsOnlineAlertComponent.razor
@@ -1,6 +1,7 @@
 ﻿@using BF2TV.Frontend.Store.FriendList
 @using BF2TV.Domain.Models.Alerts
 @using BF2TV.Frontend.Store
+@using BF2TV.Domain.Models
 @inherits FluxorComponent
 @inject IDispatcher Dispatcher
 @inject IState<AlertStore.State> State
